backend/requirements.txt
djangorestframework
django
pillow
django-cors-headers
mysqlclient